+++ client/src/components/Mainpage/MainPage.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import '../../index.css'
 
-import {withStyles} from '@material-ui/core/styles';
-import Container from '@material-ui/core/Container';
+
 
 import {Header} from "../../commons";
 import Categories from "../Categories/Categories";
@@ -10,46 +9,14 @@ import BrandsBlock from "./BrandsBlock";
 import StayInTouch from "./StayInTouch";
 import {Favorites} from "../Favorites/Favorites";
 
-const GlobalCss = withStyles({
-    // @global is handled by jss-plugin-global.
-    '@global': {
-        body: {
-            fontFamily: "'Museo Sans 500'",
-            color: "#444444"
-        },
 
-        a: {
-            textDecoration: 'none'
-        },
-       '.MuiButton-root': {
-            fontFamily: "'Museo Sans 500'",
-            background: 'linear-gradient(180deg, #6686FF 0%, #8F8DE2 100%)',
-            borderRadius: '4px',
-            border: 'none',
-            color: '#FFFFFF'
-        },
-        '.MuiTypography-root': {
-            fontFamily: "'Museo Sans 500'",
-            color: "#444444"
-        },
-        '.MuiTypography-body2': {
-            fontFamily: "'Museo Sans 500'"
-        },
-        '.MuiButtonGroup-groupedContainedPrimary:not(:last-child)': {
-          borderRight: '1px solid #EAEAEA',
-        },
-        '.MuiContainer-maxWidthMd': {
-            maxWidth: '1060px'
-        }
-    },
-})(() => null);
 
 
 
 export const MainPage = props => {
     return (
         <React.Fragment>
-            <GlobalCss/>
+
                 <Header callCenter={'1-855-324-5387'}/>
                 <BrandsBlock/>
                 <Favorites/>
